Initial Upload
This is the Mathematica notebook processing and analyzing the raw data (not included in this Git, email me if needed at jtwhite@email.wm.edu). The data here are all pertaining to GR binding DNA, in particular the canonical GRE palindrome. Also the GRE half-site, which is not discussed in the published work, as it is unfinished.